Allow subclassing by removing hard calls to URL()
One of the power of yarl is to easily create a new object based on
another.

This feature was not usable by subclasses because all methods doing this
returned a `URL` object, hard-coded.

This commit use `self.__class__` instead, to return a new object of the
same class as the original one.