URL: Add explicit_port instance property
Having access to `explicit_port` makes it easier for working with the
URL instance as a data model, like when building mutable class on top of
`yarl.URL`.

On the other hand, the existing behavior of `port` property is better
for the applications. Therefore, it's better that we keep it the way it
is.