💡Use `NO_COLOR` to override `FORCE_COLOR` @ pip
Currently, the underlying use of Rich in pip causes this library
to treat any presence of the `FORCE_COLOR` env var as a sign it's
running in a terminal, even if it's empty.

Additionally, GitHub Actions workflows don't have a syntax to unset
a globally set environment variable. This means it'll remain in the
environment and we can only unset it on the Bash script level.

Hopefully, `NO_COLOR` will just work.

Ref: https://no-color.org