Exterminate offensive references from the project
This change doesn't edit the test semantics, nor does it make any
functional changes. It removes references to the ruscist culture that
are as triggering to a lot of people as using "slave/master",
"blacklist/whitelist" terminology.

In particular, this gets rid of a link to the website that exists
to justify ethnic cleansing in Crimea and genocides in Ukraine, being
sponsored by the government of the terrorist state of muscovy.

Ref https://github.com/aio-libs/yarl/issues/955#issuecomment-1819782920