Prevent copying tmp dir to itself in build backend
Previously, setting `TMPDIR="$(pwd)/smth-inside"` would cause the
in-tree PEP 517 build backend to go into infinite recursion,
attempting to copy the project directory into a nested folder.

Specifically, this is happening in the Fedora land because macros
used in RPM packaging (`pyproject-rpm-macros`) are setting it up
like this. Additionally, this is only happening with `pip wheel` and
doesn't affect `pyproject-build`, because the latter pre-copies the
project directory by itself early, and changes current working
directory to it while the former does not.

This patch addresses the issue by excluding the temporary directory
from traversal when copying the directory tree within the build
backend.

Fixes #992.