Refactor cached_property to improve performance
stdlib cached_property takes advantage of the performance
benefits of storing the value in __dict__ which avoids the
__get__ call. This is much faster since everything can
happen in native code to fetch the value. When support
for python older than 3.12 is removed, the stdlib
cached_property can be used as versions before 3.12
had an undocumented class level lock that made the
performance profile undesirable.