Revert URL subtraction
It appears we merged the URL subtraction a bit too soon, and we need
to rework the approach. I know have ~20 hours into it trying to solve
the issues. The conclusion is that it is not ready to ship and we need
a new approach.

To ensure it does not go out before its ready, revert the original PR
and the attempts to solve the issues with it so we can start fresh.

References