Add a cache for building URLs
We currently have a cache for parsing URLs, but we did not have
one for building URLs because URL.build accepts a dict for the
query which is not hashable. Since web app tend to see the same
80% of URLs over and over, its advantageous to cache building
the URL object as aiohttp has to make them on every web request.