Improve performance of query string parsing
When looking at the performance regression in #1492,
I noticed that much of the time was spent in parse_qsl,
however we do not need all the complexity of parse_sql
since we use a very limited subset of the functionality.

We can write a faster version using the unquoter built-in
to yarl with very little code.