📦 Make release builds compile very optimized
This patch improves the in-tree build backend to allow the end-users
append `CFLAGS` and `LDFLAGS` by setting respecitve environment
variables.

It additionally sets up default compiler flags to perform build with
maximum optimization in release mode. And when line tracing is
requested, the compiler and linker flags are configured to include as
much information as possible for debugging and coverage tracking.

As a result, the C-extensions published to PyPI are going to be small
while their development/testing environment variants will be big.